requirements.txt
asgiref==3.6.0
Django==4.1.7
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.4.1
sqlparse==0.4.3
tzdata==2022.7
